+++ graphql-client/src/Data/GraphQL/Bootstrap.hs
@@ -1,6 +1,6 @@
-{- |
+{-|
 Module      :  Data.GraphQL.Bootstrap
-Maintainer  :  Brandon Chinn <brandon@leapyear.io>
+Maintainer  :  Brandon Chinn <brandonchinn178@gmail.com>
 Stability   :  experimental
 Portability :  portable